+++ data/azure_event_grid/lib/profiles/latest/modules/eventgrid_profile_module.rb
@@ -1,0 +1,89 @@
+# encoding: utf-8
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for license information.
+
+require 'azure_event_grid'
+
+module Azure::EventGrid::Profiles::Latest
+
+  module Models
+    ResourceWriteCancelData = Azure::EventGrid::V2018_01_01::Models::ResourceWriteCancelData
+    StorageBlobCreatedEventData = Azure::EventGrid::V2018_01_01::Models::StorageBlobCreatedEventData
+    ResourceDeleteSuccessData = Azure::EventGrid::V2018_01_01::Models::ResourceDeleteSuccessData
+    EventHubCaptureFileCreatedEventData = Azure::EventGrid::V2018_01_01::Models::EventHubCaptureFileCreatedEventData
+    ResourceDeleteFailureData = Azure::EventGrid::V2018_01_01::Models::ResourceDeleteFailureData
+    ResourceWriteFailureData = Azure::EventGrid::V2018_01_01::Models::ResourceWriteFailureData
+    ResourceDeleteCancelData = Azure::EventGrid::V2018_01_01::Models::ResourceDeleteCancelData
+    ResourceWriteSuccessData = Azure::EventGrid::V2018_01_01::Models::ResourceWriteSuccessData
+    EventGridEvent = Azure::EventGrid::V2018_01_01::Models::EventGridEvent
+    StorageBlobDeletedEventData = Azure::EventGrid::V2018_01_01::Models::StorageBlobDeletedEventData
+  end
+
+  #
+  # EventGridDataClass
+  #
+  class EventGridDataClass
+    attr_reader :configurable, :base_url, :options, :model_classes
+
+    def initialize(options = {})
+      if options.is_a?(Hash) && options.length == 0
+        @options = setup_default_options
+      else
+        @options = options
+      end
+
+      reset!(options)
+
+      @configurable = self
+      @base_url = options[:base_url].nil? ? nil:options[:base_url]
+      @options = options[:options].nil? ? nil:options[:options]
+
+      client_0 = Azure::EventGrid::V2018_01_01::EventGridClient.new(configurable.credentials, options)
+      if(client_0.respond_to?(:subscription_id))
+        client_0.subscription_id = configurable.subscription_id
+      end
+      add_telemetry(client_0)
+
+      @model_classes = ModelClasses.new
+    end
+
+    def add_telemetry(client)
+      profile_information = 'Profiles/Latest/EventGrid'
+      client.add_user_agent_information(profile_information)
+    end
+
+  end
+
+  class ModelClasses
+    def resource_write_cancel_data
+      Azure::EventGrid::V2018_01_01::Models::ResourceWriteCancelData
+    end
+    def storage_blob_created_event_data
+      Azure::EventGrid::V2018_01_01::Models::StorageBlobCreatedEventData
+    end
+    def resource_delete_success_data
+      Azure::EventGrid::V2018_01_01::Models::ResourceDeleteSuccessData
+    end
+    def event_hub_capture_file_created_event_data
+      Azure::EventGrid::V2018_01_01::Models::EventHubCaptureFileCreatedEventData
+    end
+    def resource_delete_failure_data
+      Azure::EventGrid::V2018_01_01::Models::ResourceDeleteFailureData
+    end
+    def resource_write_failure_data
+      Azure::EventGrid::V2018_01_01::Models::ResourceWriteFailureData
+    end
+    def resource_delete_cancel_data
+      Azure::EventGrid::V2018_01_01::Models::ResourceDeleteCancelData
+    end
+    def resource_write_success_data
+      Azure::EventGrid::V2018_01_01::Models::ResourceWriteSuccessData
+    end
+    def event_grid_event
+      Azure::EventGrid::V2018_01_01::Models::EventGridEvent
+    end
+    def storage_blob_deleted_event_data
+      Azure::EventGrid::V2018_01_01::Models::StorageBlobDeletedEventData
+    end
+  end
+end