+++ runtime/ms_rest_azure/CHANGELOG.md
@@ -1,3 +1,7 @@
+##2018.01.30 ms_rest_azure version 0.10.2
+* Added CognitiveServicesCredentials class for the cognitive services data plane SDK. Refer [PR #1257](https://github.com/Azure/azure-sdk-for-ruby/pull/1257) for further details.
+* Added TopicCredentials class for the event grid data plane SDK. Refer [PR #1257](https://github.com/Azure/azure-sdk-for-ruby/pull/1257) for further details.
+
 ##2017.12.19 ms_rest_azure version 0.10.1
 * Added support for user assigned identity to MSITokenProvider Modified portal URLs for Azure cloud environments. Refer [Issue #1175](https://github.com/Azure/azure-sdk-for-ruby/issues/1175) for further details.